New `ICE::StunPacket`: Round 1
# Details

- Replace current `RTC/StunPacket` class with a new `RTC/ICE/StunPacket` class that uses `Serializable` class.
- Move ICE related classes to `RTC/ICE/` folder.
- Create a `RTC::ICE` C++ namespace.
- Add tests for the new `RTC::ICE::StunPacket` class.

## TODO

- [ ] Remove `RTC/OLD_StunPacket.xxx` and `OLD_FuzzerStunPacket.xxx`.
- [ ] Add tests for new `RTC::ICE::StunPacket` class (so also in `meson.build`).
- [ ] Add fuzzer for new `RTC::ICE::StunPacket` class (so also in `meson.build`).

## Notes

- This PR doesn't make mediasoup use the new `RTC/StunPacket` class yet. Instead, the previous class it still used (however it's been renamed to `RTC/OLD_StunPacket.xxx` for convenience).
