+++ worker/fuzzer/include/RTC/OLD_FuzzerStunPacket.hpp
@@ -1,5 +1,5 @@
-#ifndef MS_FUZZER_RTC_STUN_PACKET_HPP
-#define MS_FUZZER_RTC_STUN_PACKET_HPP
+#ifndef MS_OLD_FUZZER_RTC_STUN_PACKET_HPP
+#define MS_OLD_FUZZER_RTC_STUN_PACKET_HPP
 
 #include "common.hpp"
