+++ worker/include/RTC/WebRtcServer.hpp
@@ -2,9 +2,9 @@
 #define MS_RTC_WEBRTC_SERVER_HPP
 
 #include "Channel/ChannelRequest.hpp"
-#include "RTC/IceCandidate.hpp"
+#include "RTC/ICE/IceCandidate.hpp"
+#include "RTC/OLD_StunPacket.hpp"
 #include "RTC/Shared.hpp"
-#include "RTC/StunPacket.hpp"
 #include "RTC/TcpConnection.hpp"
 #include "RTC/TcpServer.hpp"
 #include "RTC/TransportTuple.hpp"
@@ -57,7 +57,7 @@ namespace RTC
 	public:
 		flatbuffers::Offset<FBS::WebRtcServer::DumpResponse> FillBuffer(
 		  flatbuffers::FlatBufferBuilder& builder) const;
-		std::vector<RTC::IceCandidate> GetIceCandidates(
+		std::vector<RTC::ICE::IceCandidate> GetIceCandidates(
 		  bool enableUdp, bool enableTcp, bool preferUdp, bool preferTcp) const;
 
 		/* Methods inherited from Channel::ChannelSocket::RequestHandler. */
