+++ worker/include/RTC/WebRtcTransport.hpp
@@ -2,11 +2,11 @@
 #define MS_RTC_WEBRTC_TRANSPORT_HPP
 
 #include "RTC/DtlsTransport.hpp"
-#include "RTC/IceCandidate.hpp"
-#include "RTC/IceServer.hpp"
+#include "RTC/ICE/IceCandidate.hpp"
+#include "RTC/ICE/IceServer.hpp"
+#include "RTC/OLD_StunPacket.hpp"
 #include "RTC/Shared.hpp"
 #include "RTC/SrtpSession.hpp"
-#include "RTC/StunPacket.hpp"
 #include "RTC/TcpConnection.hpp"
 #include "RTC/TcpServer.hpp"
 #include "RTC/Transport.hpp"
@@ -20,7 +20,7 @@ namespace RTC
 	                        public RTC::UdpSocket::Listener,
 	                        public RTC::TcpServer::Listener,
 	                        public RTC::TcpConnection::Listener,
-	                        public RTC::IceServer::Listener,
+	                        public RTC::ICE::IceServer::Listener,
 	                        public RTC::DtlsTransport::Listener
 	{
 	public:
@@ -53,7 +53,7 @@ namespace RTC
 		  const std::string& id,
 		  RTC::Transport::Listener* listener,
 		  WebRtcTransportListener* webRtcTransportListener,
-		  const std::vector<RTC::IceCandidate>& iceCandidates,
+		  const std::vector<RTC::ICE::IceCandidate>& iceCandidates,
 		  const FBS::WebRtcTransport::WebRtcTransportOptions* options);
 		~WebRtcTransport() override;
 
@@ -117,22 +117,23 @@ namespace RTC
 		void OnTcpConnectionPacketReceived(
 		  RTC::TcpConnection* connection, const uint8_t* data, size_t len, size_t bufferLen) override;
 
-		/* Pure virtual methods inherited from RTC::IceServer::Listener. */
+		/* Pure virtual methods inherited from RTC::ICE::IceServer::Listener. */
 	public:
 		void OnIceServerSendStunPacket(
-		  const RTC::IceServer* iceServer,
+		  const RTC::ICE::IceServer* iceServer,
 		  const RTC::StunPacket* packet,
 		  RTC::TransportTuple* tuple) override;
 		void OnIceServerLocalUsernameFragmentAdded(
-		  const RTC::IceServer* iceServer, const std::string& usernameFragment) override;
+		  const RTC::ICE::IceServer* iceServer, const std::string& usernameFragment) override;
 		void OnIceServerLocalUsernameFragmentRemoved(
-		  const RTC::IceServer* iceServer, const std::string& usernameFragment) override;
-		void OnIceServerTupleAdded(const RTC::IceServer* iceServer, RTC::TransportTuple* tuple) override;
-		void OnIceServerTupleRemoved(const RTC::IceServer* iceServer, RTC::TransportTuple* tuple) override;
-		void OnIceServerSelectedTuple(const RTC::IceServer* iceServer, RTC::TransportTuple* tuple) override;
-		void OnIceServerConnected(const RTC::IceServer* iceServer) override;
-		void OnIceServerCompleted(const RTC::IceServer* iceServer) override;
-		void OnIceServerDisconnected(const RTC::IceServer* iceServer) override;
+		  const RTC::ICE::IceServer* iceServer, const std::string& usernameFragment) override;
+		void OnIceServerTupleAdded(const RTC::ICE::IceServer* iceServer, RTC::TransportTuple* tuple) override;
+		void OnIceServerTupleRemoved(const RTC::ICE::IceServer* iceServer, RTC::TransportTuple* tuple) override;
+		void OnIceServerSelectedTuple(
+		  const RTC::ICE::IceServer* iceServer, RTC::TransportTuple* tuple) override;
+		void OnIceServerConnected(const RTC::ICE::IceServer* iceServer) override;
+		void OnIceServerCompleted(const RTC::ICE::IceServer* iceServer) override;
+		void OnIceServerDisconnected(const RTC::ICE::IceServer* iceServer) override;
 
 		/* Pure virtual methods inherited from RTC::DtlsTransport::Listener. */
 	public:
@@ -156,7 +157,7 @@ namespace RTC
 		// Passed by argument.
 		WebRtcTransportListener* webRtcTransportListener{ nullptr };
 		// Allocated by this.
-		RTC::IceServer* iceServer{ nullptr };
+		RTC::ICE::IceServer* iceServer{ nullptr };
 		// Map of UdpSocket/TcpServer and local announced address (if any).
 		absl::flat_hash_map<RTC::UdpSocket*, std::string> udpSockets;
 		absl::flat_hash_map<RTC::TcpServer*, std::string> tcpServers;
@@ -166,7 +167,7 @@ namespace RTC
 		// Others.
 		// Whether connect() was succesfully called.
 		bool connectCalled{ false };
-		std::vector<RTC::IceCandidate> iceCandidates;
+		std::vector<RTC::ICE::IceCandidate> iceCandidates;
 		RTC::DtlsTransport::Role dtlsRole{ RTC::DtlsTransport::Role::AUTO };
 	};
 } // namespace RTC
