+++ worker/meson.build
@@ -108,8 +108,6 @@ common_sources = [
   'src/RTC/DataProducer.cpp',
   'src/RTC/DirectTransport.cpp',
   'src/RTC/DtlsTransport.cpp',
-  'src/RTC/IceCandidate.cpp',
-  'src/RTC/IceServer.cpp',
   'src/RTC/KeyFrameRequestManager.cpp',
   'src/RTC/NackGenerator.cpp',
   'src/RTC/PipeConsumer.cpp',
@@ -136,7 +134,8 @@ common_sources = [
   'src/RTC/SimpleConsumer.cpp',
   'src/RTC/SimulcastConsumer.cpp',
   'src/RTC/SrtpSession.cpp',
-  'src/RTC/StunPacket.cpp',
+  # TODO: REMOVE when new StunPacket is done.
+  'src/RTC/OLD_StunPacket.cpp',
   'src/RTC/SvcConsumer.cpp',
   'src/RTC/TcpConnection.cpp',
   'src/RTC/TcpServer.cpp',
@@ -159,6 +158,9 @@ common_sources = [
   'src/RTC/RtpDictionaries/RtpParameters.cpp',
   'src/RTC/RtpDictionaries/RtpRtxParameters.cpp',
   'src/RTC/SctpDictionaries/SctpStreamParameters.cpp',
+  'src/RTC/ICE/IceCandidate.cpp',
+  'src/RTC/ICE/IceServer.cpp',
+  'src/RTC/ICE/StunPacket.cpp',
   'src/RTC/RTP/Packet.cpp',
   'src/RTC/RTP/ProbationGenerator.cpp',
   'src/RTC/RTP/SharedPacket.cpp',
@@ -633,7 +635,8 @@ fuzzer_sources = common_sources + [
   'fuzzer/src/RTC/FuzzerRtpRetransmissionBuffer.cpp',
   'fuzzer/src/RTC/FuzzerRtpStreamSend.cpp',
   'fuzzer/src/RTC/FuzzerSeqManager.cpp',
-  'fuzzer/src/RTC/FuzzerStunPacket.cpp',
+  # TODO: REMOVE when new StunPacket is done.
+  'fuzzer/src/RTC/OLD_FuzzerStunPacket.cpp',
   'fuzzer/src/RTC/FuzzerTrendCalculator.cpp',
   'fuzzer/src/RTC/RTP/FuzzerPacket.cpp',
   'fuzzer/src/RTC/RTP/FuzzerProbationGenerator.cpp',
