+++ worker/src/RTC/ICE/StunPacket.cpp
@@ -1,0 +1,94 @@
+#define MS_CLASS "RTC::ICE::StunPacket"
+// #define MS_LOG_DEV_LEVEL 3
+
+#include "RTC/ICE/StunPacket.hpp"
+#include "Logger.hpp"
+#include "MediaSoupErrors.hpp"
+#include "Utils.hpp"
+#include <cstdio> // std::snprintf()
+
+namespace RTC
+{
+	namespace ICE
+	{
+		/* Class variables. */
+
+		const uint8_t StunPacket::MagicCookie[] = { 0x21, 0x12, 0xA4, 0x42 };
+
+		/* Class methods. */
+
+		bool StunPacket::IsStun(const uint8_t* buffer, size_t bufferLength)
+		{
+			// clang-format off
+			return (
+				// STUN headers are 20 bytes.
+				(bufferLength >= 20) &&
+				// @see RFC 7983.
+				(buffer[0] < 3) &&
+				// Magic cookie must match.
+				(buffer[4] == StunPacket::MagicCookie[0]) && (buffer[5] == StunPacket::MagicCookie[1]) &&
+				(buffer[6] == StunPacket::MagicCookie[2]) && (buffer[7] == StunPacket::MagicCookie[3])
+			);
+			// clang-format on
+		}
+
+		StunPacket* StunPacket::Parse(const uint8_t* buffer, size_t bufferLength)
+		{
+			MS_TRACE();
+
+			if (!StunPacket::IsStun(buffer, bufferLength))
+			{
+				MS_WARN_TAG(ice, "not a STUN Packet");
+
+				return nullptr;
+			}
+
+			auto* packet = new StunPacket(const_cast<uint8_t*>(buffer), bufferLength);
+
+			// `bufferLength` must be the exact length of the Packet, so let's assign
+			// it immediately.
+			packet->SetLength(bufferLength);
+
+			// TODO
+		}
+
+		/* Instance methods. */
+
+		StunPacket::StunPacket(uint8_t* buffer, size_t bufferLength)
+		  : Serializable(buffer, bufferLength)
+		{
+			MS_TRACE();
+
+			// TODO
+		}
+
+		StunPacket::~StunPacket()
+		{
+			MS_TRACE();
+		}
+
+		void StunPacket::Dump(int indentation) const
+		{
+			MS_TRACE();
+
+			MS_DUMP_CLEAN(indentation, "<ICE::StunPacket>");
+
+			// TODO
+
+			MS_DUMP_CLEAN(indentation, "<ICE::StunPacket>");
+		}
+
+		StunPacket* StunPacket::Clone(uint8_t* buffer, size_t bufferLength) const
+		{
+			MS_TRACE();
+
+			auto* clonedPacket = new StunPacket(buffer, bufferLength);
+
+			Serializable::CloneInto(clonedPacket);
+
+			// TODO: Copy pointers and so on.
+
+			return clonedPacket;
+		}
+	} // namespace ICE
+} // namespace RTC
