+++ worker/src/RTC/OLD_StunPacket.cpp
@@ -1,7 +1,7 @@
 #define MS_CLASS "RTC::StunPacket"
 // #define MS_LOG_DEV_LEVEL 3
 
-#include "RTC/StunPacket.hpp"
+#include "RTC/OLD_StunPacket.hpp"
 #include "Logger.hpp"
 #include "MediaSoupErrors.hpp"
 #include "Utils.hpp"
